Output metadata type and route (#1394)
* Add OutputMetadataResponse type

* Fix return of ConsumedOutput::milestone_timestamp

* Fix TODOs

* Remove useless milestone_timestamp as u32 casts

* Add output metadata route

* Add create_output_metadata

+++ bee-message/src/output/unlock_condition/mod.rs
@@ -242,7 +242,7 @@ impl UnlockConditions {
         milestone_timestamp: u32,
     ) -> &'a Address {
         self.expiration()
-            .and_then(|e| e.return_address_expired(milestone_index, milestone_timestamp as u32))
+            .and_then(|e| e.return_address_expired(milestone_index, milestone_timestamp))
             .unwrap_or(address)
     }
 }
